+++ src/index.ts
@@ -1,3 +1,4 @@
 export const MARKORBIT_CORE_VERSION = '0.0.0';
 
 export * from './domains/index.ts';
+export * from './objects/index.ts';